feat: Make Sidebar And Release Card Component reusable
