+++ frontend/css/main.less
@@ -27,6 +27,7 @@
 @import "search-track.less";
 @import "stats.less";
 @import "switch.less";
+@import "sidebar.less";
 @import "new-navbar.less";
 @import "tags.less";
 @import "stats-art-creator.less";
